+++ packages/@sanity/vision/package.config.ts
@@ -10,9 +10,14 @@ export default defineConfig({
     treeshake: {moduleSideEffects: true},
     output: {
       intro: (chunkInfo) => {
-        if (chunkInfo.isEntry && chunkInfo.name === 'index') {
-          return `import './bundle.css'`
-        }
+        /**
+         * TODO: we are avoiding importing the bundle.css file here because it's producing
+         * errors when using `sanity` with node or for server rendering
+         * `Error: Unknown file extension ".css"`
+         */
+        // if (chunkInfo.isEntry && chunkInfo.name === 'index') {
+        //   return `import './bundle.css'`
+        // }
 
         return ''
       },